Assignment-3/pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "adeelhashmat", email = "adeelhashmatkhan@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.1.0",
]
[tool.uv.scripts]
run_agent = "python studybuddy.py"

[project.scripts]
studybuddy = "studybuddy:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
